feat(client/src): Input Media Streams Display
