+++ client/src/components/Landing.js
@@ -1,6 +1,4 @@
 import '../App.css';
-import { useState, useRef } from 'react';
-import firebase from '../firebase/index';
 import { Link } from "react-router-dom";
 //asets
 import LandingVector from '../assets/LandingVector.png';
@@ -9,121 +7,12 @@ import { Row, Col, Typography, Button, Input, Divider } from 'antd';
 import { PlusOutlined, RiseOutlined, FieldNumberOutlined } from '@ant-design/icons';
 
 
-function Landing() {
-  
-  const [roomId, setRoomId] = useState("");
-
-  const localStream = useRef();
-  const remoteStream = useRef();
+function Landing(props) {
 
   const { Title } = Typography;
 
-  const configuration = {
-    iceServers: [
-      {
-        urls: [
-          'stun:stun1.l.google.com:19302',
-          'stun:stun2.l.google.com:19302',
-        ],
-      },
-    ],
-    iceCandidatePoolSize: 10,
-  };
-
-  const registerPeerConnectionListeners = () =>{
-    console.log(peerConnection)
-    peerConnection.addEventListener('icegatheringstatechange', () => {
-      console.log(
-          `ICE gathering state changed: ${peerConnection.iceGatheringState}`);
-    });
-  
-    peerConnection.addEventListener('connectionstatechange', () => {
-      console.log(`Connection state change: ${peerConnection.connectionState}`);
-    });
-  
-    peerConnection.addEventListener('signalingstatechange', () => {
-      console.log(`Signaling state change: ${peerConnection.signalingState}`);
-    });
-  
-    peerConnection.addEventListener('iceconnectionstatechange ', () => {
-      console.log(
-          `ICE connection state change: ${peerConnection.iceConnectionState}`);
-    });
-  }
-
-  let peerConnection = null;
-
   const newMeetingHandler = async() =>{
-    const db = firebase.firestore();
-    const roomRef = await db.collection('rooms').doc();
-    console.log('Create PeerConnection with configuration: ', configuration);
-    const stream = await navigator.mediaDevices.getUserMedia({video: true, audio: true});
-    localStream.current = stream;
-    remoteStream.current = new MediaStream();
-    peerConnection = new RTCPeerConnection(configuration); 
-    registerPeerConnectionListeners();
-    localStream.current.getTracks().forEach(track => {
-        peerConnection.addTrack(track, localStream.current);
-      });
-    // Code for collecting ICE candidates below
-    const callerCandidatesCollection = roomRef.collection('callerCandidates');
-
-    peerConnection.addEventListener('icecandidate', event => {
-        if (!event.candidate) {
-          console.log('Got final candidate!');
-          return;
-        }
-        console.log('Got candidate: ', event.candidate);
-        callerCandidatesCollection.add(event.candidate.toJSON());
-      });
-    // Code for collecting ICE candidates above
-
-    // Code for creating a room below
-    const offer = await peerConnection.createOffer();
-    await peerConnection.setLocalDescription(offer);
-    console.log('Created offer:', offer);
-
-    const roomWithOffer = {
-        'offer': {
-        type: offer.type,
-        sdp: offer.sdp,
-        },
-    };
-    await roomRef.set(roomWithOffer);
-    setRoomId(roomRef.id);
-    // Code for creating a room above
-
-    peerConnection.addEventListener('track', event => {
-      console.log('Got remote track:', event.streams[0]);
-      event.streams[0].getTracks().forEach(track => {
-        console.log('Add a track to the remoteStream:', track);
-        remoteStream.current.addTrack(track);
-      });
-    });
-
-    // Listening for remote session description below
-    roomRef.onSnapshot(async snapshot => {
-      const data = snapshot.data();
-      if (!peerConnection.currentRemoteDescription && data && data.answer) {
-        console.log('Got remote description: ', data.answer);
-        const rtcSessionDescription = new RTCSessionDescription(data.answer);
-        await peerConnection.setRemoteDescription(rtcSessionDescription);
-      }
-    });
-    // Listening for remote session description above
-
-    // Listen for remote ICE candidates below
-    roomRef.collection('calleeCandidates').onSnapshot(snapshot => {
-      snapshot.docChanges().forEach(async change => {
-        if (change.type === 'added') {
-          let data = change.doc.data();
-          console.log(`Got new remote ICE candidate: ${JSON.stringify(data)}`);
-          await peerConnection.addIceCandidate(new RTCIceCandidate(data));
-        }
-      });
-    });
-    // Listen for remote ICE candidates above
-    console.log(roomId);
+    props.setOperation("create");
   }
 
   return (
